Final version before review

+++ src/transform_funcs.c
@@ -6,7 +6,7 @@
 /*   By: sdiez-ga <sdiez-ga@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/18 17:17:40 by sdiez-ga          #+#    #+#             */
-/*   Updated: 2022/07/20 18:21:22 by sdiez-ga         ###   ########.fr       */
+/*   Updated: 2022/10/04 16:09:14 by sdiez-ga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ t_complex	*transform_coordinates(t_complex *coords, t_vars *vars)
 t_complex	apply_transform(t_complex coords, t_vars *vars)
 {
 	coords.re = ((double)(coords.re * 5.0) / vars->width) - \
-				(double)vars->offset_x;
+				vars->offset_x;
 	coords.im = ((double)(coords.im * 3.0) / vars->height) - \
-				(double)vars->offset_y;
+				vars->offset_y;
 	coords.re *= vars->scale;
 	coords.im *= vars->scale;
 	return (coords);
